Extract the driver's DAG-walk utilities into IncrementalWalks
The paged-epoch symbol walks and their per-node set cache, the capped
DAG-size counts with the big-formula memo, the multi-root union count
and the budget-bounded reachability probe are one self-contained bundle
of walk machinery whose only tie to the driver is epoch lifetime. They
now live in IncrementalWalks.h (header-only; its single manager
dependency, the first node number pages are offset from, arrives
through the constructor), and the driver keeps one-line forwarders so
sixty-odd call sites read as they always have. The epoch hooks --
reclaim the symbol sets on a SAT-only rebuild, release everything at a
relief rotation -- became named methods, so the rotation's release list
shrinks by the three entries it used to have to keep in step by hand.
